+++ scripts/events/signup-flow-event.js
